Custom galaxies now support infrastructure
- Added support for infrastructure in custom galaxies.
- Moved custom galaxy related carrier code from gameCreate.ts to
custom.ts.
- Added a check for bad numeral values in star and carrier properties to
gameCreateValidation.ts.

Removed unary plus operators preceding id, playerId, wormholeToStarId
and specialistId fields - these made their _checkStarProperty() checks
non-functional, as any invalid input would be converted into NaN,
which is of type 'number'. Changed id, playerId, wormholeToStarId and
orbiting fields to be of 'string' type instead of 'number' type (the
values are replaced with actual DB IDs, so their type does not actually
matter). Setting it to 'string' allows the user to use a JSON from an
existing game. Values (of these fields) of type 'number' (such as ones
in generated JSON from current galaxy editors) are changed to type
'string' via toString().

_checkNumericalProperty() in gameCreateValidation.ts prevents NaN and
Infinity from being used in fields of type 'number', as well as checking
for non-whole numbers in certain fields. It returns false for undefined
or null fields/values, but does not throw an error in those cases, as
certain fields are optional. This check is currently only used for
fields of stars and carriers, but it may make sense to use it for game
settings as well (since a user may provide bad values).

The 1500 star limit check now also applies for games with a custom
galaxy.